rename to kri
I decided that the current name (yocto) will cause problemw should this text editor ever become succesful. There is another software project with yocto as name: Yocto Linux. Although I did know it existed when I named this text editor, I liked the name and never thought it might be capable to compete with well-known text editors.
There is a goat is Crete called "kri-kri".  I cut it to kri to be short. It is short and memorable, plus no other software has that name. Although KRI stands for something else, it shouldn't pose a problem as it is unrelated to software.